scrapper/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
pymongo==4.10.1
pydantic==2.8.2
httpx==0.27.2
beautifulsoup4==4.12.3
lxml==5.2.2
APScheduler==3.10.4
